test: add hookLists tests, reorganize drupal integration tests
