+++ tests/integration/frameworks/drupal/11.1/test_hook_implementations_map_empty_key.php
@@ -1,11 +1,12 @@
 <?php
+
 /*
  * Copyright 2020 New Relic Corporation. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
 
 /*DESCRIPTION
-Verify agent behavior when key value is an empty array
+Verify agent behavior when key is an empty string
 */
 
 /*SKIPIF
@@ -26,15 +27,15 @@ newrelic.framework = drupal8
 /*EXPECT
 */
 
-require_once __DIR__ . '/mock_module_handler_empty_array.php';
+require_once __DIR__ . '/mock_module_handler_empty_key.php';
 
 // This specific API is needed for us to instrument the ModuleHandler
 class Drupal
 {
-  public function moduleHandler()
-  {
-    return new Drupal\Core\Extension\ModuleHandler();
-  }
+    public function moduleHandler()
+    {
+        return new Drupal\Core\Extension\ModuleHandler();
+    }
 }
 
 // Create module handler
