+++ tests/integration/frameworks/drupal/11.3/test_hooklists_invalid_map.php
@@ -1,0 +1,43 @@
+<?php
+
+/*
+ * Copyright 2020 New Relic Corporation. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/*DESCRIPTION
+Verify the agent does not crash when handling an invalid hookLists property.
+*/
+
+/*SKIPIF
+<?php
+if (version_compare(PHP_VERSION, '8.3', '<')) {
+    die("skip: PHP >= 7.4 required\n");
+}
+*/
+
+/*INI
+newrelic.framework = drupal8
+*/
+
+/*EXPECT_TRACED_ERRORS null */
+
+/*EXPECT_ERROR_EVENTS null */
+
+/*EXPECT
+*/
+
+require_once __DIR__ . '/mock_module_handler_invalid_map.php';
+
+// This specific API is needed for us to instrument the ModuleHandler
+class Drupal
+{
+    public function moduleHandler()
+    {
+        return new Drupal\Core\Extension\ModuleHandler();
+    }
+}
+
+// Create module handler
+$drupal = new Drupal();
+$handler = $drupal->moduleHandler();
